disable suspending the wifi mmc port for now
It seems the bcmsdh_sdmmc driver does not provide suspend / resume
callbacks making  mmc_suspend_hosts return the -ENOSYS (-38) error.

Why this error does not happen on all devices is mystery.

It seems the only possible way to control the powermanagement
currently available is via the wlarm_le utility.

We should revisit this later, when have the most recent sources for
the dhd driver.