Reenable suspending mmc ports but keep the wifi port from being powered down
Why we are customer 5 now, I don't know, so disable the check.
Also add KEEP_POWER to the wifi mmc host to keep it from being powered down
after its suspend function is called - the wifi should be able to wakeup the
system, but we want to call the dhd suspend method to be able to hopefully
do a little bit of power saving.