Fix hanging powerdown
On some msp430 ntx-firmware versions the poweroff does not succeed
in special undetermined conditions.

This results in the device not powering down but beeing stuck in
the poweroff loop. Therefore if after 10 tries the system is still
not powered down, we try to reset it instead.

If the reset succeeds, it will bring the device into a consistent
state at least and not leave it hanging where the user does not
now what happened.