elan_i2c: Fix read hello packet when coming out of deep sleep
The touch controller gets powered down somewhere else when the system
enters the deep sleep and gets powered on again on resume.

Of course the hello packet has to be read again after this powercycle.

This fixes the problem, that the interrupt line stays 0 after sleep until
the first touch happens.