Add early mac initialization
dhd_bus_start will call dhd_read_macaddr that should get the MAC
address from the chip.

I've moved the bus start to dhd_init but I have also kept it on
dhd_open, because I'am not sure if the init is only called on driver
initialization and open every time the hw is detected.