zforce_ts: Always emit wakeup events on touch events
We do not emit wakeup events for interrupts (command responses)
we receive when suspending - as it would keep the system awake
indefinitly. But we always want to emit wakeup events when
a touch event is received, as this means the user touched the
screen.

This should improve the doubletouch issue down the road.