add second reset mechanism
On the E60672 there are situations where the msp430 based powermanagement
unit is in a confused state and not able to do the 10 second shutdown.

But as our investigations showed, it seems to be still able to accept
normal i2c commands.

Therefore implement a delayed work, that gets executed when the power button
is pressed for 11 seconds. And issue a reset command to the msp.