reenable the powerbutton handling also in the notifier
We moved the reenabling to an early resume stage, to catch all early button
presses. But this leaves out the symmetry of the suspend process in case of
failure. When the suspend fails, all steps done to suspend are undone.

So if the suspend fails directly behind the pm_notifier, only the notifier
itself would be undone. And currently the powerbutton would stay locked,
as the reenabling is done at a vastly earlier stage.

Therefore reenable the button handling also in the notifier.