zforce_ts: improve reset handling
The zforce issues the bootcomplete notification when the boot is complete
after reset was pulled high. So there is no need to wait specific amounts
of time, but simpy for the notification. This also requires the irq-line
to always stay enabled, to be able to receive out-of-band notifications.

Furthermore put the reset-handling into a delayed work, as restarting the
ic from the interrupt context produces a deadlock, as the interrupt-
handling is waiting for an interrupt (boot notification) to happen, which
of course can't be processed, as the handler is waiting.