move the delayed rtc read into the generic power-unlock conditional
Previously the rtc read was called when the up-event was reported.
When returning from sleep there is not powerkey up-event reported
to keep the system from immediately entering sleep again. Therefore
the rtc read also was not executed when returning from sleep.

Now in the unlock-stage it will be really executed everytime also
including when returning from sleep.